WCW-5162: Support permissions other than active (#16)
Oof, I didn't realize that the `make format` wasn't run against the last PR from the WAX team. Going to merge this in and run formatting against the rest of the code so its not an issue moving forward. Will release as 1.5.1.

I also didn't catch the hardcoded permission - so thanks for fixing that!